Issue #1: Rename trello api module
